pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["dbpal"]
namespaces = true

[project]
name = "dbpal"
authors = [
  {name = "Michael Chow", email = "mc_al_github@fastmail.com"}
]
description = ""
requires-python = ">=3.8"
dependencies = [
  "plum-dispatch",
  "fsspec",
  "gcsfs",
  "importlib_resources",
  "pandas",
  "sqlalchemy",
  "sqlalchemy-bigquery",
  "siuba",
]
version = "0.0.1"

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-dotenv",
  "duckdb_engine",
  "pandas-gbq"
]

[tool.pytest.ini_options]
testpaths = ["dbpal"]
